requirements.txt
ultralytics
pymodbus==3.6.8
opencv-python
customtkinter
pillow
numpy